librmnetctl: Fix incorrect handling of messages with return type data
When an message with return type data fails in kernel, the error
is handled incorrectly.

Consider a scenario when we try to query the vnd name corresponding
to a non existent vnd id.

/ # rmnetcli getvndname 25
rmnetcli getvndname 25
LIBRARY ERROR: Return type is invalid

We appear to be getting a library error even though the actual
error originates from the kernel. The actual error for this case
should be a kernel error and the return code should be Invalid
request / Unsupported scenario.

Fix this by checking for return codes in case there is an error
when a return data type fails.

Change-Id: Ideb765f786cf19e29f3f496a6c8d7da92e769111