Fix WiFi on user builds
E android.hardware.wifi@1.0-service: Failed to open wlan fw path param: Permission denied

The problem here isn't that we didn't have the permission, we did but that it was being
set too late and so causing the system to not have the permission. By moving it up, the
system can now set the permision for wifi and fw can be written and wifi starts properly.

Change-Id: I3f5a88a45b9aead0435562cae12f993f1cff2e84